backend/requirements.txt
fastapi
uvicorn[standard]
python-dotenv
qdrant-client
langchain-text-splitters
sentence-transformers
google-genai